New endpoint to support android dashboard move getSpreadsheetValues and spreadsheetId's to utilities file; create android route
